+++ src/plugins/signoff/constants.js
@@ -1,0 +1,4 @@
+export const PLUGIN_REVIEW_REQUEST = "PLUGIN_REVIEW_REQUEST";
+export const PLUGIN_DECLINE_REQUEST = "PLUGIN_DECLINE_REQUEST";
+export const PLUGIN_SIGNOFF_REQUEST = "PLUGIN_SIGNOFF_REQUEST";
+export const SIGNOFF_WORKFLOW_INFO = "SIGNOFF_WORKFLOW_INFO";
